authorization implemented

+++ src/api/notes/handler.js
@@ -18,11 +18,16 @@ class NotesHandler {
       const {
         title = 'untitled', body, tags
       } = request.payload;
+      const {
+        id: credentialId
+      } = request.auth.credentials;
       const noteId = await this._service.addNote({
         title,
         body,
-        tags
+        tags,
+        owner: credentialId,
       });
+
       const response = h.response({
         status: 'success',
         message: 'Catatan berhasil ditambahkan',
@@ -54,8 +59,11 @@ class NotesHandler {
     }
   }
 
-  async getNotesHandler() {
-    const notes = await this._service.getNotes();
+  async getNotesHandler(request) {
+    const {
+      id: credentialId
+    } = request.auth.credentials;
+    const notes = await this._service.getNotes(credentialId);
     return {
       status: 'success',
       data: {
@@ -69,7 +77,13 @@ class NotesHandler {
       const {
         id
       } = request.params;
+      const {
+        id: credentialId
+      } = request.auth.credentials;
+
+      await this._service.verifyNoteOwner(id, credentialId);
       const note = await this._service.getNoteById(id);
+
       return {
         status: 'success',
         data: {
@@ -103,7 +117,10 @@ class NotesHandler {
       const {
         id
       } = request.params;
-
+      const {
+        id: credentialId
+      } = request.auth.credentials;
+      await this._service.verifyNoteOwner(id, credentialId);
       await this._service.editNoteById(id, request.payload);
 
       return {
@@ -136,6 +153,10 @@ class NotesHandler {
       const {
         id
       } = request.params;
+      const {
+        id: credentialId
+      } = request.auth.credentials;
+      await this._service.verifyNoteOwner(id, credentialId);
       await this._service.deleteNoteById(id);
 
       return {
